Added logic to mongoose.connect() to the HEROKU database via `process.env.MONGODB_URI` IF the variable exists.
IF the environment variable does not exist, we connect directly to our localhost string.

GOING FORWARD: I would like you to set up a local environment variable called `MONGODB_URI` and connect using it
in localhost as well.

What this should look like is a single line of code that looks like this:

		mongoose.connect(process.env.MONGODB_URI);

...instead of the `if` statement we are currently using.